Add option for CSV output
Right now this is not so different from the initial list format, because we don't export any other data.

It should eventually be a relatively helpful option once we do start exporting other data though (widely understood by a number of programs).

Note that the old option, now called 'token-list', remains the default (for now).